requirements.txt
molo.core<6.0.0,>=5.0.0
celery<4.0
django-celery
psycopg2
html5lib==0.9999999
fcm-django
